requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.1.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
lxml>=4.9.0
fpdf2>=2.7.0
selenium>=4.8.0
webdriver-manager>=4.0.0
anthropic>=0.30.0
python-dotenv>=1.0.0
